config: move DDNS warn validators to own file
Pure code-motion for #4845. Move the DDNS WARN validators
(validateDDNSBackendWarnings, validateSurfaceADDNSWarnings), the ddns*
parsing/mirror helpers, the ddnsKnownDyndns2NameSet table, and the
hasFamily helper (used only by validateSurfaceADDNSWarnings) verbatim from
compiler_validate_warn.go into compiler_validate_warn_ddns.go.

Bodies are byte-identical; only the file each declaration lives in and the
per-file import block change. Same package config, same signatures. No
behavior change.
